Fix the core problem, by storing resources on tiles as references to those resources instead of embedded data.
This uses keys (which for now are the same as names) for resources, similar to the terrain keys.  I'm not 100% sure if keys are needed, but my thinking is (1) names being the identifier is a large part of why Civ scenarios are hard to make multilingual, and (2) using keys as the references can be done consistently, so you always know that key means "reference to something".