Switch to the Babylon background.
Even with some cropping from the original 5:4, the best I can do is about 16:11.75, very slightly wider than 4:3.  This will have to do, it's impossible not to lose the Tower with more cropping.

On the plus side, the default startup res is 4:3, so it works well there.

Also saw that you're supported to commit the .import files, so I'll do that now.