Set the font color for the main menu labels to 90% black, and move them to the left.
Moving them left fixes the jumping around when you resize, and looks better with the Babylon background.

At some point maybe we can get them properly anchored to the right, but that conflicts with the clouds in Babylon's background, and I'm a bit worn out for figuring that out tonight.