Explorer AI now prioritizes tiles that will reveal more tile knowledge.
At the beginning, this looks great, as AI unit fan out to discover new tiles.

But at some point the flaw - that it's always looking for the local maximum - reveals itself, as units get trapped in corners of the map.

Probably going to refactor things a bit before fixing that.