Work-around to loading images via res://SomePath.jpg not working in release mode.  Unfortunately, I couldn't find a good way to keep that working in debug mode; most didn't work, and the rest will be deprecated with Godot 4.0 and added complexity.
So instead, I added a utility method that will use the current method in debug mode, and use a release-appropriate method.  That latter method was discovered in a comment in the Godot Mono source code; unfortunately I closed the file before noting which one it was.

Also, commit a change to the VSStub.csproj project that enables building it in release mode (see answer #2 at https://stackoverflow.com/questions/15134384/the-outputpath-property-is-not-set-for-project).  This might fix the issues that WildWeazel and others have had with VSStub.