Refactor unit AIs into their own folder.  Make sure they all return a value for whether they succeeded.  And add a loop in PlayerAI so if a unit fails at its task, it gets a new one... but not infinitely.
Currently, once an AI player has explored its entire landmass, it throws the warning about failing twice.  This is because the AI gets the Explore task again by default, and is expected for now.  Now that it's verified that is working correctly, I'll add logic to ensure units don't get Explorer tasks if there isn't anything left to explore.

Also, the change in SettlerAI to improve logging is because the "could not get part of the path" showed up in the logs for me.  Unfortunately I was in Run instead of Debug mode at the time.

Looking forward to that proper logging framework!