Cities now produce food, shields, and theoretically commerce based on tile yields.
The downside is tile yields are low now, with no irrigation or resource contributions.  Unfortunate for any civ that starts in an area that can never reach size 3.

But it's still more realistic.  Production only based on the 8 neighboring tiles and city center for now.

NOTE: This is a slightly modified re-commit on a new branch.  There were too many conflicts in TurnHandling.cs for non-manual merging.