Re-implement SeaUnit as a "Category" for a unit.
This will allow extension in the future for new categories, and also works better with the JSON than the multiple C# types of UnitPrototypes.

Not necessarily the final design, but solves the land/sea split for now in an extensible way (we still need to fix movement bugs and train the AI with naval units).