Add a StrategicPriority structure, and a first StrategicPriority (ExpansionPriority) that follows that structure.
Once another priority or two are added, the next step will be having the AI choose a priority.  This will be the first "arbiter" added to the AI, at least with a planned structure.  CityLocationAI kind of does arbitrate, but doesn't have a generic, pluggable architecture for that.  StrategicPriorityAI will.