Start adding code leading towards being able to save strategic AI data on the player object.
Right now we basically duplicate data between the engine and the data layers.  I'm highly tempted to put StrategicPriority.cs in the data layer, while leaving the implementations in the engine.  This gets rid of the data duplication and is more object-oriented; we also already have methods on some of the data objects.

Comparing this to the unit AI, I see that for those, I have a different storage structure for each data storage object - SettlerAI stores different things than ExplorerAI, for example.  Whereas for the strategic priorities, I've structured it so that the data will be stored in a simple property map (or dictionary in C# parlance).  This is a more flexible structure, and lends itself towards this extensibility.  On the other hand, I see why I wanted more concrete types for unit AI - it stores UnitPath and Tile objects, for example.

Hmmmmm.  MapUnit has a UnitAIData object, and UnitAIData is in the data project.  But what if the concrete type implementing UnitAIData were in a different project?  It should still serialize correctly, right?  That could give us both the flexibility we need in storing various types, and the save structure we need, assuming the implementations handle serialization adequately.

I think that might all work.