Move StrategicPriority into the data layer, to avoid data duplication.
Now, there's a complexifier, but it's going to be a complexifier regardless of where that class lives.  Serialization.

Sources:
 - https://docs.microsoft.com/en-us/dotnet/standard/serialization/system-text-json-polymorphism
 -https://docs.microsoft.com/en-us/dotnet/standard/serialization/system-text-json-converters-how-to?pivots=dotnet-6-0#support-polymorphic-deserialization

Basically, as long as all the properties are on the base class, things will work.  But if there are properties defined on implementing classes, they won't get serialized without a custom serializer.  Le sigh.  It's the same in Java, at least with Jackson.

Java's GSON library (the superior serialization library) offers a solution to this: https://jansipke.nl/serialize-and-deserialize-a-list-of-polymorphic-objects-with-gson/  Won't help with C7 directly, but maybe the concept could.

However, according to https://docs.microsoft.com/en-us/dotnet/standard/serialization/system-text-json-migrate-from-newtonsoft-how-to?WT.mc_id=DOP-MVP-4039670&pivots=dotnet-6-0, Newtonsoft JSON *does* support polymorphic serialization and deserialization in C#.  So maybe we should use it.