Add multiple priorities.  The AI will choose 2-4 priorities, depending on whether the prioritization ranks are close together (choose more), or whether a few are obviously more important (choose fewer).
This isn't really applied yet, but my vision for it is that the AI will give the heaviest bonus weighting to its top priority, and lesser weighting to its secondary/tertiary/etc. priorities.  For example, let's say that WarPriority gives a 400% bonus to the likelihood of building military units.  As a top priority, that full bonus would be applied.  As a secondary, half of it would.  As a tertiary, a fourth of it would.  It would still make its preference more likely, but less so if it's a secondary or lower priority.

(Long-term, it would also be cool to have the priorities be able to loop into future AI components, such as a military advisor style one.  WarPriority should always make military units more likely, but if the AI is actively losing cities, the bonus should be even higher)