The AI will now make a semi-dynamic calculation of what to build.
This commit mostly just includes the structure, and right now it seems to always result in building Galleys.  This is because (1) there's no randomization, (2) it doesn't consider whether cities are landlocked, and (3) the base scoring doesn't include being able to build cities, so Settlers always have a negative score.  Also, that there isn't yet a component to say, "we already have enough galleys".

So it doesn't really make things better yet (unless you like lots of galleys), but it adds some structure and shines some light on some of the challenges.