#267 - Add Serilog to C7Engine, and replace all Console logs with it in that project.godot
Used what I thought was the appropriate level.  We can shift them as time goes; I expect we'll generally be demoting logs from info to debug (or verbose) as more is added.

Also set the default minimum level to "info".  We should be selectively enabling debug for certain areas, which I plan to test next.