More investigation.  Jerry-rigged the random number generator to always use the same seed so I can compare like results.  Weirdly I had to set the seed right before random start loc generation, for some reason setting it up front wasn't sticking, even though I verified it hit that line with breakpoints and that it generated the same random # on the evaluate expression console there.  Oh well.  That was becoming a distraction and the main goal here isn't making the seed settable, so the work-around works.
So far, I've figured out that the logging actually takes a lot of time.  At the end of turn 4, America sends a Warrior to 21, 41.  It takes 2-3 ms with lines 126-130 of ExplorerAI - the detailed logging - commented out, and 15-16 ms with those detailed logs enabled.  That is six lines of logs taking about 13ms, or 2ms per line.

That might not always be most of the time, but it is sometimes.

Also properly set namespaces.