This commit illustrates that logging stuff is slow.  At first I switched to using Serilog templates as recommended (make sure you don't @ the thing you are logging or it'll serialize it).  But now I'm just logging a hard-coded string, and it still takes 11-12 ms for the loop, versus 2 ms without it.  Logging six hard-coded lines takes 9-10 milliseconds, or at least 1.5 milliseconds per log statement.
This can also be observed by the timestamps on the logs.

This is a problem as it means logging is slow, and logging to figure out why something is slow will make it significantly more slow.  We could still log around one thing to see if it's slow, but if we try to log with much granularity, we'll rapidly slow things down.

Why is it slow?  That is still TBD.  Maybe Serilog is slow.  Maybe our Godot sink is slow.  Maybe something in our formatting or config is slow.  Maybe it's something specific to my console that is slow.