Add file sink, use it only instead of Godot sink.
The catapult example average 3 to 3.2 seconds very consistently, including with custom format, for our Catapult 6-tile evaluation example with 6 logs.  The best-ever, no-formatting Godot time was 6 ms; no logs was 1 ms.

In a more complex Warrior evaluating 23 tile example, no logs is 8 ms, no sinks and file logs are both 10 ms, Godot sink is 59 ms.

This is probably acceptable, the downside being it's slightly inconvenient having to tail a file, but it will allow evaluating the actual algorithm, not the logging algorithm.