Think I've got a good way to shorten the exploration time.  It was looking at all possible tiles that would allow continued exploration, and seeing how far away they were.  That took a while, especially because pathing takes longer the farther away a tile is.
Now, we sort by crow-flies distance first.  Then we look at how long it takes to get there, starting with the closest-as-the-crow-flies one, and skip checking any whose crow-flies distance is greater than the verified actually-pathable distance we've already found.  So if the closest crow-flies is blocked by water, we'll check more, but most of the time we quit checking very early and only check the closest tile(s).

In the Babylon Warrior turn 12 example, this takes the time down from 86ms to 1ms - very noticeable to not noticeable at all.