Avoid having to set the seed in multiple places.
Figured out that passing it as a parameter was somehow affecting the seed/next number it generated.  Not sure why, I wouldn't have guessed that based on the API, but making it static and not passing it as a parameter was what was necessary to get consistent results from the same seed without setting it any time it was received as a parameter.