Basically making logs around city production better.  There's an edge case where it builds a Settler in a size-2 city which causes a crash, and I'm trying to figure out why.  For now I've added an error log and a work-around so it won't crash, but it's a problem.  Thought I finally had the logs in place, but the log file hit the 5 MB limit and stopped recording - I thought it'd just start over or roll off the old ones, but no, it stops altogether.
I think the city production work was concurrent with Serilog, which is why a lot of it used Console.log.

Aside from the size 2 settler bug, the AI *is* settling in CxC.  Had to tone down barbs since they don't know to protect their settlers or hunt barbs yet, but once I did that, it works.  The AI probably puts too much emphasis on resource location (at least early on) and not enough on closeness, but I can see the appeal of the locations it settles.