Use the imported data (and remember to call the import method!).  Also inflate it from the JSON.
This fixes 06 MP Age of Discovery.biq, but unfortunately doesn't yet work with the static JSON, because the references are to the full list of PRTOs, and we are only importing a select list.

I think this shows that the limiter hack, which we added so we didn't have all units available before technology is added, is now costing as much to maintain as it is saving.  I plan to move the limiter to the what-can-you-build check instead of the unit-prototype-import next, to fix this problem.