Display custom scenario graphics, so long as there's only one item in the scenario search path.
I'd thought we had this before, but it was only in the legacy Civ3Map, and only when you tossed a string for the folder in the source code.

Now, it stores the search path in Util, and the Civ3MediaPath method uses that automatically.  From a pure OO standpoint, the old way of having to pass in the search path as an argument may be cleaner, but we already were calling that method from 25 places and it just didn't seem to make sense to have that cognitive overhead in 25 places when it's only going to change when a new game/scenario is loaded.