Add a TNT layer that shows the bonus grassland overlay.  For now I haven't added other TNT support, as we don't have a "Show Food and Shields on Map" equivalent.
Also worth noting, for some reason in .SAV files, non-grassland tiles show up as having bonus yield.  I've excluded these visually, but confirmed it's not just out C7 file but happens with a Civ .SAV as well.  The SAV-reading code looks correct to me, so it's a mystery... perhaps Civ3 really does set the bonus flag for non-grasslands when it generates a random map.