370 - Add Kyriakos's Persian city graphics.
Since these are a different size (600x547) than the base ones (501x380), a code update is also needed.  Most of the extra size is pink space, so I think it's mostly just giving it a bigger possible canvas.

Checking with the Manhattan graphics, arbitrary size city graphics are supported in Civ3, so we should support that as well.  We might need a flag for high-res when we get to that, but can cross that bridge when it's time to do so.