172 - Start extracting city drawing into a scene.  Right now it just draws the base city graphics and not the label, but (almost surprisingly) it works.
How much of a benefit is it versus the old style?  Not sure yet, it probably will be more as it increases.  I could also see it helping if we need something temporary such as fireworks or WLTKD to be added, which the CityLayer could handle without the CityScene having to worry about it.  And the CityScene could remember which graphics is uses which would be nice, we already had issues with having to cache city color backgrounds.