#172 - Attempt to convert to using one texture (and maybe some string drawing) that gets drawn on to the screen at once, to avoid the residual drawing items problem of the previous commit.
There's an ERROR: Drawing is only allowed inside NOTIFICATION_DRAW, _draw() function or 'draw' signal. that happens when I try to draw strings even inside the _draw signal, which is unfortunate.  Not sure why yet, but we'll have to figure it out to go this route.

Aside from that and figuring out how to compose the city graphic and label textures, it was relatively easy to get the city graphic showing in the appropriate location with margins, and it does stick around in the right place when panning.