#172 - Clean up CityScene, use proper children for everything.
Realized after some investigation that the problem with the label showing up twice when the growth time changes from 10 to 9 is not because of something in CityScene, but because CityLayer adds duplicate CityScene instances every time it renders.  This is a carryover from when we were just blitting the graphics onto the scene all the time, but a new model is needed for that.