#172 - CityLayer no longer adds infinite CityScenes.
This is now functionally equivalent to the old code.  I am not sure whether it performs quite as well.  Oh and it just started spewing out ERROR: Disconnecting nonexistent signal 'changed', slot: 37634:_emit_theme_changed., not sure what that is about but probably something with how I did the themes for the fonts.  Seed is 882942484 for reproducing it.