Add turn time logging.
This isn't directly related to the city scene but I was doing some testing to see if performance was about the same (it seems to be) and this seemed like a good way to add some more objective measurements.  Though to be fair it doesn't really measure graphics performance, more so turn processing.