Support great wonders that add buildings to other cities
I spent a while considering how to implement this, since it seems like there are two ways:

1. Add the wonder-provided cities to each city, which is what this PR does.

2. Any time we need to look at the buildings in a city, check all of the other cities for great wonders and see if they would add a building to our city.

Option 2 seemed worse to me, and would have to be done carefully to avoid some annoying recursion. Option 1 also makes it easy to track where a building came from, which would allow us to easily implement the "you can sell a granary if you built it before you got the pyramids" logic of the original game.

The downside to option 1 is some duplication of logic between the game runtime and the import logic, but it doesn't seem excessive. Let me know if you disagree.